import fix, partial error module...
The major change here is that import ... from lodash doesn't play nicely. That was creating a global window  object and loading the ENTIRE lodash lib. Changing this cut the App size in half. I'll be hunting for more ways to cut it down as it's still ~1.7mb, which feels too big.

Also the humble beginings of an actual error handler. This will in most cases activate an error modal screen that lets the user know what went wrong, and very rarely try to fix itself... Mostly because most errors are going to be connection issues, or (hopefully never) api refactoring issues that break the app. Either way, the user should know what's happening so they can let us know about it.