Error polling for reconnection.
There were some hefty changes in this commit, mostly for generic error handling. Now, whenever an error is thrown, the app trys to automatically reconnect to the internet. The error module, and haltState implementation now provides a logical, generic approach for exiting an error state: if on any successful response haltState is set to true, emit setHaltState event to turn off. Various modules listen for this switch and behave appropriately depending on the state.