App / pokeList state saving.
This was kind of complicated, mostly because I wanted to get it right and
keeps things as decoupled as possible...

At the top level, the app now tracks the correct pokedex entry number and
contains an additional `id` variable. THe difference between the two are
unique pokemon id and relative pokedex entry number.

The PokeList now participates in a little call/response from the top level app
to check whether it starts at zero, or has a state to resume. Since everything
is rather singular, there's no need to update any other models.