breaking free from the state prison.
storing, accessing and maintaining state was holding this app hostage. I've implemented a centralized model using vuex to solve it.

Most of the code is relatively terse and self-explanatory. There are still some kinks to be worked out, but now shared state between components exist in one location for a much more sane app strucutre.